Move integration methods to another file
